+++ src/constants/regex.js
@@ -1,0 +1,1 @@
+export const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
